[Refactor] axiosData 함수 이동
